+++ jdkjson-examples/src/main/java/org/jdkstack/jdkjson/examples/Examples.java
@@ -5,7 +5,7 @@ import java.io.FileInputStream;
 import java.io.InputStreamReader;
 import java.nio.charset.StandardCharsets;
 import java.util.Random;
-import org.jdkstack.jdkjson.core.JsonReaderV2;
+import org.jdkstack.jdkjson.core.JsonReaderV1;
 
 public class Examples {
 
@@ -43,10 +43,10 @@ public class Examples {
           "{\"age\":"
               + new Random().nextInt(80)
               + ",\"clothes\":{\"coat\":\"Nike\",\"trousers\":\"adidas\",\"shoes\":\"安踏\"},\"friends\":[{\"age\":24,\"clothes\":{\"coat\":\"Nike\",\"trousers\":\"adidas\",\"shoes\":\"安踏\"},\"fullName\":{},\"hobbies\":[\"篮球\",\"游泳\",\"coding\"],\"name\":\"小明\"},{\"age\":24,\"clothes\":{\"coat\":\"Nike\",\"trousers\":\"adidas\",\"shoes\":\"安踏\"},\"fullName\":{},\"hobbies\":[\"篮球\",\"游泳\",\"coding\"],\"name\":\"Tony\"},{\"age\":24,\"clothes\":{\"coat\":\"Nike\",\"trousers\":\"adidas\",\"shoes\":\"安踏\"},\"fullName\":{},\"hobbies\":[\"篮球\",\"游泳\",\"coding\"],\"name\":\"陈小二\"}],\"fullName\":{},\"hobbies\":[\"篮球\",\"游泳\",\"coding\"],\"name\":\"邵同学\"}";
-      JsonReaderV2 jsonParserV2 = new JsonReaderV2(map);
-      jsonParserV2.deserializeLru();
+      // JsonReaderV2 jsonParserV2 = new JsonReaderV2(map);
+      // jsonParserV2.deserializeLru();
 
-      // JsonReaderV1.deserializeLru(map);
+      JsonReaderV1.deserializeLru(map);
     }
 
     /*    for (int i = 0; i < 2000000; i++) {
